requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.4
Pillow==9.3.0
pycodestyle==2.10.0
sqlparse==0.4.3
tzdata==2022.7
